Uppercase the first letter of the product name in 'create-products.js'.
The faker.hacker.noun() method returns a string with a lowercase
first letter. As I'm using this as the first word of the product
name, I want that letter to be uppercase. It seems like there should
be a standard JS method for this, but there doesn't seem to be.