Renamed str -> hackerNoun.
I'm trying to make sure all of my variable names describe accurately
what is inside of the variable. 'str' was too broad a description in
this case, as the variable holds an uppercased version of the noun
returned from faker.hacker.noun() I thought hackerNoun was a more
descriptive name.