Add 'bash' to code sections of README and add '$' character before each command.
I think it is clearer to have the '$' character in front of
commands in the README that are supposed to be run in the
terminal. Even though I hate having to delete the character
in the terminal when I copy and paste. I may go back and remove
these later.