Move the logic for validating an incoming :id param into validate.js.
I wanted to create a function in validate.js that could be used
as a callback to the router.param('id') middleware, and stack
different callbacks in there like I did in the router.post('/)
middleware. Unfortunately, router.param() only allows one callback.
I tried calling it multiple times, but didn't like the way that
looked, so I moved the :id validation logic into the validate.js
file and cleaned up the if() statement at the begininng of
router.param('id') a little.