Fixed logic for handling errors.
OK, so revelation here. When you call next(err) it DOES NOT return
from the function you are in. It passes the error to the error
handler, then function execution resumes from the point that it
left off in the route-handler. This means, that anytime you
call next(err), you need to be sure that you are guarding any
further logic with either if-blocks or a return statement.