Refactor GET /products to use if() conditionals differently.
I was using an if..else conditional to determine the response, but
this will get messy once I have to add in the num_results query,
since it has to be applied AFTER any keyword search. It seems
logical to have ONE return object, res.locals.products, that can
have different filters applied to it by query string params. So,
instead of returning one object in the if(), and another in the
else(), the if() simply makes modifications to the returned object
and there is no need for an else().